requirements.txt
colorama
Flask
html5lib
pandas
pytest
python-dotenv
requests
tabulate
yfinance
scipy
logger
